feat: add general parsing function
Added file named general-parsing.h and general-parsing.c that has the function to generally parse the data in csv file. To use do, parsing(input, n, output1, output2, output3, .., outputn);
where input is the result of fgets from csv (a line), n is the amount of data to get from the input, and output<k> for storing data from the result of parsing.