feat: add functions to insert by id and to free list to ListObat and ListPenyakit ADT
Function insert by id is to insert an item that has an definite id to the list with indeks = id. The nEff component will adjust to be after the value of id if nEff <= id

Function free is to free the memory for the list when the list it's not used anymore. Useful when exiting the program.